requirements.txt
uvicorn~=0.30.5
fastapi~=0.112.0
SQLAlchemy~=2.0.32
pydantic~=2.8.2
APScheduler~=3.10.4
python-dotenv~=1.0.1
pymysql~=1.1.1
celery~=5.4.0